+++ lib/src/components/effects/sound_effect_sheet.dart
@@ -8,8 +8,11 @@ import 'package:zego_uikit/zego_uikit.dart';
 // Project imports:
 import 'package:zego_uikit_prebuilt_live_streaming/src/components/effects/effect_grid.dart';
 import 'package:zego_uikit_prebuilt_live_streaming/src/internal/internal.dart';
+import 'package:zego_uikit_prebuilt_live_streaming/src/live_streaming_translation.dart';
 
 class ZegoSoundEffectSheet extends StatefulWidget {
+  final ZegoTranslationText translationText;
+
   final List<VoiceChangerType> voiceChangerEffect;
   final ValueNotifier<String> voiceChangerSelectedIDNotifier;
 
@@ -18,6 +21,7 @@ class ZegoSoundEffectSheet extends StatefulWidget {
 
   const ZegoSoundEffectSheet({
     Key? key,
+    required this.translationText,
     required this.voiceChangerEffect,
     required this.voiceChangerSelectedIDNotifier,
     required this.reverbEffect,
@@ -86,7 +90,7 @@ class _ZegoSoundEffectSheetState extends State<ZegoSoundEffectSheet> {
           ),
           SizedBox(width: 10.r),
           Text(
-            'Audio effect',
+            widget.translationText.audioEffectTitle,
             style: TextStyle(
               fontSize: 36.0.r,
               color: const Color(0xffffffff),
@@ -109,7 +113,7 @@ class _ZegoSoundEffectSheetState extends State<ZegoSoundEffectSheet> {
     }
 
     voiceChangerModel = ZegoEffectGridModel(
-      title: 'Voice changing',
+      title: widget.translationText.audioEffectVoiceChangingTitle,
       selectedID: widget.voiceChangerSelectedIDNotifier,
       items: voiceChangerEffect
           .map(
@@ -124,7 +128,7 @@ class _ZegoSoundEffectSheetState extends State<ZegoSoundEffectSheet> {
                 icon: PrebuiltLiveStreamingImage.asset(
                     'assets/icons/voice_changer_${effect.name}_selected.png'),
               ),
-              iconText: effect.text,
+              iconText: voiceChangerTypeText(effect),
               onPressed: () {
                 ZegoUIKit().setVoiceChangerType(effect.key);
               },
@@ -148,7 +152,7 @@ class _ZegoSoundEffectSheetState extends State<ZegoSoundEffectSheet> {
     }
 
     reverbPresetModel = ZegoEffectGridModel(
-      title: 'Reverb',
+      title: widget.translationText.audioEffectReverbTitle,
       selectedID: widget.reverbSelectedIDNotifier,
       items: reverbEffect
           .map(
@@ -159,7 +163,7 @@ class _ZegoSoundEffectSheetState extends State<ZegoSoundEffectSheet> {
                 icon: PrebuiltLiveStreamingImage.asset(
                     'assets/icons/reverb_preset_${effect.name}.png'),
               ),
-              iconText: effect.text,
+              iconText: reverbTypeText(effect),
               onPressed: () {
                 ZegoUIKit().setReverbType(effect.key);
               },
@@ -168,10 +172,71 @@ class _ZegoSoundEffectSheetState extends State<ZegoSoundEffectSheet> {
           .toList(),
     );
   }
+
+  String voiceChangerTypeText(VoiceChangerType voiceChangerType) {
+    switch (voiceChangerType) {
+      case VoiceChangerType.none:
+        return widget.translationText.voiceChangerNoneTitle;
+      case VoiceChangerType.littleBoy:
+        return widget.translationText.voiceChangerLittleBoyTitle;
+      case VoiceChangerType.littleGirl:
+        return widget.translationText.voiceChangerLittleGirlTitle;
+      case VoiceChangerType.deep:
+        return widget.translationText.voiceChangerDeepTitle;
+      case VoiceChangerType.crystalClear:
+        return widget.translationText.voiceChangerCrystalClearTitle;
+      case VoiceChangerType.robot:
+        return widget.translationText.voiceChangerRobotTitle;
+      case VoiceChangerType.ethereal:
+        return widget.translationText.voiceChangerEtherealTitle;
+      case VoiceChangerType.female:
+        return widget.translationText.voiceChangerFemaleTitle;
+      case VoiceChangerType.male:
+        return widget.translationText.voiceChangerMaleTitle;
+      case VoiceChangerType.optimusPrime:
+        return widget.translationText.voiceChangerOptimusPrimeTitle;
+      case VoiceChangerType.cMajor:
+        return widget.translationText.voiceChangerCMajorTitle;
+      case VoiceChangerType.aMajor:
+        return widget.translationText.voiceChangerAMajorTitle;
+      case VoiceChangerType.harmonicMinor:
+        return widget.translationText.voiceChangerHarmonicMinorTitle;
+    }
+  }
+
+  String reverbTypeText(ReverbType reverbType) {
+    switch (reverbType) {
+      case ReverbType.none:
+        return widget.translationText.reverbTypeNoneTitle;
+      case ReverbType.ktv:
+        return widget.translationText.reverbTypeKTVTitle;
+      case ReverbType.hall:
+        return widget.translationText.reverbTypeHallTitle;
+      case ReverbType.concert:
+        return widget.translationText.reverbTypeConcertTitle;
+      case ReverbType.rock:
+        return widget.translationText.reverbTypeRockTitle;
+      case ReverbType.smallRoom:
+        return widget.translationText.reverbTypeSmallRoomTitle;
+      case ReverbType.largeRoom:
+        return widget.translationText.reverbTypeLargeRoomTitle;
+      case ReverbType.valley:
+        return widget.translationText.reverbTypeValleyTitle;
+      case ReverbType.recordingStudio:
+        return widget.translationText.reverbTypeRecordingStudioTitle;
+      case ReverbType.basement:
+        return widget.translationText.reverbTypeBasementTitle;
+      case ReverbType.popular:
+        return widget.translationText.reverbTypePopularTitle;
+      case ReverbType.gramophone:
+        return widget.translationText.reverbTypeGramophoneTitle;
+    }
+  }
 }
 
 void showSoundEffectSheet(
   BuildContext context, {
+  required ZegoTranslationText translationText,
   required List<VoiceChangerType> voiceChangeEffect,
   required List<ReverbType> reverbEffect,
   required ValueNotifier<String> voiceChangerSelectedIDNotifier,
@@ -198,6 +263,7 @@ void showSoundEffectSheet(
           child: SizedBox(
             height: 600.r,
             child: ZegoSoundEffectSheet(
+              translationText: translationText,
               voiceChangerEffect: voiceChangeEffect,
               voiceChangerSelectedIDNotifier: voiceChangerSelectedIDNotifier,
               reverbEffect: reverbEffect,