+++ lib/src/live_streaming_translation.dart
@@ -25,11 +25,54 @@ class ZegoTranslationText {
   String requestCoHostFailedToast;
   String repeatInviteCoHostFailedToast;
 
+  String messageEmptyToast;
+
   ZegoDialogInfo cameraPermissionSettingDialogInfo;
   ZegoDialogInfo microphonePermissionSettingDialogInfo;
   ZegoDialogInfo receivedCoHostRequestDialogInfo;
   ZegoDialogInfo receivedCoHostInvitationDialogInfo;
   ZegoDialogInfo endConnectionDialogInfo;
+
+  /// effect
+  String audioEffectTitle;
+  String audioEffectReverbTitle;
+  String audioEffectVoiceChangingTitle;
+
+  /// voice effect
+  String voiceChangerNoneTitle;
+  String voiceChangerLittleBoyTitle;
+  String voiceChangerLittleGirlTitle;
+  String voiceChangerDeepTitle;
+  String voiceChangerCrystalClearTitle;
+  String voiceChangerRobotTitle;
+  String voiceChangerEtherealTitle;
+  String voiceChangerFemaleTitle;
+  String voiceChangerMaleTitle;
+  String voiceChangerOptimusPrimeTitle;
+  String voiceChangerCMajorTitle;
+  String voiceChangerAMajorTitle;
+  String voiceChangerHarmonicMinorTitle;
+
+  /// revert effect
+  String reverbTypeNoneTitle;
+  String reverbTypeKTVTitle;
+  String reverbTypeHallTitle;
+  String reverbTypeConcertTitle;
+  String reverbTypeRockTitle;
+  String reverbTypeSmallRoomTitle;
+  String reverbTypeLargeRoomTitle;
+  String reverbTypeValleyTitle;
+  String reverbTypeRecordingStudioTitle;
+  String reverbTypeBasementTitle;
+  String reverbTypePopularTitle;
+  String reverbTypeGramophoneTitle;
+
+  /// beauty effect
+  String beautyEffectTypeWhitenTitle;
+  String beautyEffectTypeRosyTitle;
+  String beautyEffectTypeSmoothTitle;
+  String beautyEffectTypeSharpenTitle;
+  String beautyEffectTypeNoneTitle;
 
   ZegoTranslationText({
     String? disagreeButton,
@@ -50,11 +93,45 @@ class ZegoTranslationText {
     String? audienceRejectInvitationToast,
     String? requestCoHostFailedToast,
     String? repeatInviteCoHostFailedToast,
+    String? messageEmptyToast,
     ZegoDialogInfo? cameraPermissionSettingDialogInfo,
     ZegoDialogInfo? microphonePermissionSettingDialogInfo,
     ZegoDialogInfo? receivedCoHostRequestDialogInfo,
     ZegoDialogInfo? receivedCoHostInvitationDialogInfo,
     ZegoDialogInfo? endConnectionDialogInfo,
+    String? audioEffectTitle,
+    String? audioEffectReverbTitle,
+    String? audioEffectVoiceChangingTitle,
+    String? voiceChangerNoneTitle,
+    String? voiceChangerLittleBoyTitle,
+    String? voiceChangerLittleGirlTitle,
+    String? voiceChangerDeepTitle,
+    String? voiceChangerCrystalClearTitle,
+    String? voiceChangerRobotTitle,
+    String? voiceChangerEtherealTitle,
+    String? voiceChangerFemaleTitle,
+    String? voiceChangerMaleTitle,
+    String? voiceChangerOptimusPrimeTitle,
+    String? voiceChangerCMajorTitle,
+    String? voiceChangerAMajorTitle,
+    String? voiceChangerHarmonicMinorTitle,
+    String? reverbTypeNoneTitle,
+    String? reverbTypeKTVTitle,
+    String? reverbTypeHallTitle,
+    String? reverbTypeConcertTitle,
+    String? reverbTypeRockTitle,
+    String? reverbTypeSmallRoomTitle,
+    String? reverbTypeLargeRoomTitle,
+    String? reverbTypeValleyTitle,
+    String? reverbTypeRecordingStudioTitle,
+    String? reverbTypeBasementTitle,
+    String? reverbTypePopularTitle,
+    String? reverbTypeGramophoneTitle,
+    String? beautyEffectTypeWhitenTitle,
+    String? beautyEffectTypeRosyTitle,
+    String? beautyEffectTypeSmoothTitle,
+    String? beautyEffectTypeSharpenTitle,
+    String? beautyEffectTypeNoneTitle,
   })  : disagreeButton = disagreeButton ?? 'Disagree',
         agreeButton = agreeButton ?? 'Agree',
         startLiveStreamingButton = startLiveStreamingButton ?? 'Start',
@@ -77,6 +154,7 @@ class ZegoTranslationText {
             'Failed to connect with the co-host, please try again.',
         repeatInviteCoHostFailedToast = repeatInviteCoHostFailedToast ??
             "You've sent the invitation, please wait for confirmation.",
+        messageEmptyToast = messageEmptyToast ?? 'Say something...',
         audienceRejectInvitationToast =
             audienceRejectInvitationToast ?? '%0 refused to be a co-host.',
         requestCoHostFailedToast =
@@ -116,5 +194,47 @@ class ZegoTranslationText {
             ZegoDialogInfo(
               title: 'End the connection',
               message: 'Do you want to end the cohosting?',
-            );
+            ),
+        audioEffectTitle = audioEffectTitle ?? 'Audio effect',
+        audioEffectReverbTitle = audioEffectReverbTitle ?? 'Reverb',
+        audioEffectVoiceChangingTitle =
+            audioEffectVoiceChangingTitle ?? 'Voice changing',
+        voiceChangerNoneTitle = voiceChangerNoneTitle ?? 'None',
+        voiceChangerLittleBoyTitle = voiceChangerLittleBoyTitle ?? 'Little boy',
+        voiceChangerLittleGirlTitle =
+            voiceChangerLittleGirlTitle ?? 'Little girl',
+        voiceChangerDeepTitle = voiceChangerDeepTitle ?? 'Deep',
+        voiceChangerCrystalClearTitle =
+            voiceChangerCrystalClearTitle ?? 'Crystal-clear',
+        voiceChangerRobotTitle = voiceChangerRobotTitle ?? 'Robot',
+        voiceChangerEtherealTitle = voiceChangerEtherealTitle ?? 'Ethereal',
+        voiceChangerFemaleTitle = voiceChangerFemaleTitle ?? 'Female',
+        voiceChangerMaleTitle = voiceChangerMaleTitle ?? 'Male',
+        voiceChangerOptimusPrimeTitle =
+            voiceChangerOptimusPrimeTitle ?? 'Optimus Prime',
+        voiceChangerCMajorTitle = voiceChangerCMajorTitle ?? 'C major',
+        voiceChangerAMajorTitle = voiceChangerAMajorTitle ?? 'A major',
+        voiceChangerHarmonicMinorTitle =
+            voiceChangerHarmonicMinorTitle ?? 'Harmonic minor',
+        reverbTypeNoneTitle = reverbTypeNoneTitle ?? 'None',
+        reverbTypeKTVTitle = reverbTypeKTVTitle ?? 'Karaoke',
+        reverbTypeHallTitle = reverbTypeHallTitle ?? 'Hall',
+        reverbTypeConcertTitle = reverbTypeConcertTitle ?? 'Concert',
+        reverbTypeRockTitle = reverbTypeRockTitle ?? 'Rock',
+        reverbTypeSmallRoomTitle = reverbTypeSmallRoomTitle ?? 'Small room',
+        reverbTypeLargeRoomTitle = reverbTypeLargeRoomTitle ?? 'Large room',
+        reverbTypeValleyTitle = reverbTypeValleyTitle ?? 'Valley',
+        reverbTypeRecordingStudioTitle =
+            reverbTypeRecordingStudioTitle ?? 'Recording studio',
+        reverbTypeBasementTitle = reverbTypeBasementTitle ?? 'Basement',
+        reverbTypePopularTitle = reverbTypePopularTitle ?? 'Pop',
+        reverbTypeGramophoneTitle = reverbTypeGramophoneTitle ?? 'Gramophone',
+        beautyEffectTypeWhitenTitle =
+            beautyEffectTypeWhitenTitle ?? 'Skin Tone',
+        beautyEffectTypeRosyTitle = beautyEffectTypeRosyTitle ?? 'Blusher',
+        beautyEffectTypeSmoothTitle =
+            beautyEffectTypeSmoothTitle ?? 'Smoothing',
+        beautyEffectTypeSharpenTitle =
+            beautyEffectTypeSharpenTitle ?? 'Sharpening',
+        beautyEffectTypeNoneTitle = beautyEffectTypeNoneTitle ?? 'None';
 }